{Print ellipses for unprinted line and column changes}
{!
Use ellipses to indicate unprinted lines and columns

If the first line isn't printed, an ellipsis line (a line consisted
of only an ellipsis) is printed at the beginning; if the last line
isn't printed, an ellipsis line is printed at the end; if two line
changes are separated by more than 2 * `contextLines` lines, some
lines between them aren't printed, and an ellipsis line is printed
in their place.

If the first column of a changed line isn't printed, an ellipsis
is printed at the beginning of the first column change; if the
last column of a changed line isn't printed, an ellipsis is printed
at the end of the last column change; if two column changes in a
changed line are separated by more than 2 * `contextColumns` columns,
some columns between them aren't printed, and two ellipses are printed
instead: one at the end of the first column change, and the other
at the beginning of the second column change.

Additionally, if some line or column changes are not printed due to
the limitation by `maxChunks` and `maxChunksPerLine`, extra ellipses
are used to notify the user of this.

For column changes in a changed line, the first `colRes` entry that
isn't printed will have its line and column numbers printed (in cyan),
followed by an ellipsis (rather than its content, in default color).
This tells the user there are column changes not printed, starting
at this one.

For line changes, the first line change (a '-', '+', or '*') that
isn't printed will have its line number printed, followed by an
ellipsis, and the entire line is styled according to its nature
(by default red for '-', green for '+', cyan for '*'). This tells
the user there are line changes not printed, starting at this one.
!}